+++ IsekaiMod/Config/ModSupport.cs
@@ -46,6 +46,7 @@ namespace IsekaiMod.Utilities {
                     AutoRime.Add();
                     AutoBurning.Add();
                     AutoFlaring.Add();
+                    AutoPiercing.Add();
                 }
                 if (IsMysticalMayhemEnabled()) {
                     Main.Log("Mystical Mayhem 0.1.5 Support Enabled.");