+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 ### Added
-- Added `AutoRime` OP ability (requires TabletopTweaks-Base).
-- Added `AutoBurning` OP ability (requires TabletopTweaks-Base).
-- Added `AutoFlaring` OP ability (requires TabletopTweaks-Base).
+- Added `Auto Rime` OP ability (requires TabletopTweaks-Base).
+- Added `Auto Burning` OP ability (requires TabletopTweaks-Base).
+- Added `Auto Flaring` OP ability (requires TabletopTweaks-Base).
+- Added `Auto Piercing` OP ability (requires TabletopTweaks-Base).
 - Updated mod support for ExpandedContent.
 ### Fixed
 - Fixed missing dialogue for drow ambush encounter.